fix(css): fix error
